+++ app/components/recording_studio_presskits/press_kits/show_component.rb
@@ -3,7 +3,8 @@
 module RecordingStudioPresskits
   module PressKits
     class ShowComponent < ViewComponent::Base
-      def initialize(press_kit_recording:, section_recordings:, picker_types:, add_path:, remove_path:, reorder_path:) # rubocop:disable Metrics/ParameterLists
+      def initialize(press_kit_recording:, section_recordings:, picker_types:, add_path:, remove_path:, reorder_path:, # rubocop:disable Metrics/ParameterLists
+                     preview_path: nil, public_path: nil, publish_path: nil)
         super()
         @press_kit_recording = press_kit_recording
         @section_recordings = section_recordings
@@ -11,6 +12,13 @@ module RecordingStudioPresskits
         @add_path = add_path
         @remove_path = remove_path
         @reorder_path = reorder_path
+        @preview_path = preview_path
+        @public_path = public_path
+        @publish_path = publish_path
+      end
+
+      def live?
+        @press_kit_recording.respond_to?(:currently_published?) && @press_kit_recording.currently_published?
       end
 
       def kit_title