+++ lib/recording_studio_presskits/admin.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
-require "recording_studio_presskits/admin/press_kits_list_widget"
+require "recording_studio_presskits/admin/press_kits_published_widget"
+require "recording_studio_presskits/admin/press_kits_unpublished_widget"
 require "recording_studio_presskits/admin/press_kits_section"
 
 module RecordingStudioPresskits
@@ -9,7 +10,8 @@ module RecordingStudioPresskits
       def register!
         return unless defined?(::RecordingStudioAdmin)
 
-        RecordingStudioAdmin.register_widget(PressKitsListWidget)
+        RecordingStudioAdmin.register_widget(PressKitsPublishedWidget)
+        RecordingStudioAdmin.register_widget(PressKitsUnpublishedWidget)
         RecordingStudioAdmin.register_section(PressKitsSection)
       end
     end