+++ test/recording_studio_presskits_test.rb
@@ -4,7 +4,7 @@ require "test_helper"
 
 class RecordingStudioPresskitsTest < Minitest::Test
   def test_version_matches_release
-    assert_equal "0.5.0", ::RecordingStudioPresskits::VERSION
+    assert_equal "0.6.0", ::RecordingStudioPresskits::VERSION
   end
 
   def test_engine_exists
@@ -21,7 +21,7 @@ class RecordingStudioPresskitsTest < Minitest::Test
     assert_includes gemspec, 'spec.add_dependency "recording_studio_trashable", "~> 0.4"'
     assert_includes gemspec, 'spec.add_dependency "recording_studio_duplicatable", "~> 0.4"'
     assert_includes gemspec, 'spec.add_dependency "flat_pack", ">= 0.1.133"'
-    refute_includes gemspec, 'spec.add_dependency "recording_studio_publishable"'
+    assert_includes gemspec, 'spec.add_dependency "recording_studio_publishable", "~> 0.2"'
     refute_includes gemspec, 'spec.add_dependency "recording_studio_attachable"'
     refute_includes gemspec, 'spec.add_dependency "recording_studio_api"'
   end
@@ -37,6 +37,7 @@ class RecordingStudioPresskitsTest < Minitest::Test
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_orderable", tag: "0.2.0"'
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_trashable", tag: "0.4.0"'
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_duplicatable", tag: "0.4.0"'
+    assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_publishable", tag: "v0.2.0"'
     refute_includes gemfile, "recording_studio/v3.0.0"
     refute_includes gemfile, 'tag: "v0.6.0"'
     refute_includes gemfile, 'tag: "v0.1.134"'
@@ -61,6 +62,9 @@ class RecordingStudioPresskitsTest < Minitest::Test
     assert_includes source, "include RecordingStudio::Capabilities::Orderable.to"
     assert_includes source, "include RecordingStudio::Capabilities::Trashable.to"
     assert_includes source, "RecordingStudio::Capabilities::Duplicatable.to"
+    assert_includes source, "RecordingStudio::Capabilities::Publishable.to"
+    assert_includes source, 'public_controller: "recording_studio_presskits/public_press_kits"'
+    assert_includes source, "public_action: :show"
     assert_includes source, 'suffix: " (Copy)"'
     assert_includes source, "exclude_children: []"
     refute_includes source, "include_children: true"
@@ -80,7 +84,7 @@ class RecordingStudioPresskitsTest < Minitest::Test
     assert_includes source, "def press_kit_type_name"
     assert_includes source, "def picker_types"
     assert_includes source, "RecordingStudio.recordable_type_name"
-    assert_includes source, "RecordingStudio.allowed_parent_types_for"
+    assert_includes source, "RecordingStudio.declared_allowed_parent_types_for"
     refute_includes source, "Block"
     refute_includes source, "Slot"
   end
@@ -111,8 +115,8 @@ class RecordingStudioPresskitsTest < Minitest::Test
     assert_includes routes, 'mount RecordingStudioTrashable::Engine, at: "/recording_studio_trashable"'
     assert_includes routes, 'mount RecordingStudioDuplicatable::Engine, at: "/recording_studio_duplicatable"'
     assert_includes routes, 'mount RecordingStudioPresskits::Engine, at: "/recording_studio_presskits"'
+    assert_includes routes, 'mount RecordingStudioPublishable::Engine, at: "/"'
     assert_includes routes, 'recording_studio_admin_for :admin, at: "/admin", root_section: :press_kits'
-    refute_includes routes, "recording_studio_publishable"
   end
 
   def test_dummy_login_layout_keeps_flatpack_assets_without_tight_main_offset
@@ -140,6 +144,7 @@ class RecordingStudioPresskitsTest < Minitest::Test
     assert_includes sources_task, '"recording_studio_duplicatable"'
     assert_includes sources_task, '"recording_studio_admin"'
     assert_includes sources_task, '"recording_studio_presskits"'
+    assert_includes sources_task, '"recording_studio_publishable"'
     refute_includes tailwind_source, "@theme"
     refute_includes tailwind_source, ":root {"
     refute_includes tailwind_source, "--color-fp-primary"
@@ -153,6 +158,7 @@ class RecordingStudioPresskitsTest < Minitest::Test
     assert_includes initializer_source, '"RecordingStudioPresskits::PressKit"'
     assert_includes initializer_source, '"FakeBlock"'
     assert_includes initializer_source, '"AdminRoot"'
+    assert_includes initializer_source, '"RecordingStudioPublishable::Publishable"'
     refute_includes initializer_source, "config.include_children"
     refute_includes initializer_source, "config.features."
     refute_includes initializer_source, "v3"
@@ -178,6 +184,8 @@ class RecordingStudioPresskitsTest < Minitest::Test
     assert_includes readme, "tag: \"2.0.0\""
     assert_includes readme, "tag: \"0.2.0\""
     assert_includes readme, "tag: \"0.4.0\""
+    assert_includes readme, "tag: \"v0.2.0\""
+    assert_includes readme, "PressKit.indexable"
     assert_includes readme, "Press kit"
     assert_includes readme, "RecordingStudioPresskits::PressKit"
     assert_includes readme, "recording_studio_orderable_reorder!"
@@ -221,6 +229,7 @@ class RecordingStudioPresskitsTest < Minitest::Test
     assert_includes source, "include RecordingStudio::Capabilities::Trashable.to"
     refute_includes source, "Capabilities::Orderable"
     refute_includes source, "Capabilities::Duplicatable"
+    refute_includes source, "Capabilities::Publishable"
     refute_includes source, ".with("
     refute_includes source, ".enabled"
   end
@@ -240,20 +249,31 @@ class RecordingStudioPresskitsTest < Minitest::Test
 
   def test_engine_registers_admin_list_section
     section = File.read(File.expand_path("../lib/recording_studio_presskits/admin/press_kits_section.rb", __dir__))
-    widget = File.read(File.expand_path("../lib/recording_studio_presskits/admin/press_kits_list_widget.rb", __dir__))
+    published = File.read(
+      File.expand_path("../lib/recording_studio_presskits/admin/press_kits_published_widget.rb", __dir__)
+    )
+    unpublished = File.read(
+      File.expand_path("../lib/recording_studio_presskits/admin/press_kits_unpublished_widget.rb", __dir__)
+    )
 
     assert_includes section, 'key "press_kits"'
-    assert_includes section, 'widget "widgets.press_kits.list"'
-    refute_includes section, "published"
-    refute_includes widget, "type :number"
-    assert_includes widget, "type :list"
-    assert_includes widget, "KitQuery.live_kits"
-    refute_includes widget, "Publishable"
+    assert_includes section, 'widget "widgets.press_kits.published"'
+    assert_includes section, 'widget "widgets.press_kits.unpublished"'
+    refute_includes section, "widgets.press_kits.list"
+    refute_includes published, "type :number"
+    refute_includes unpublished, "type :number"
+    assert_includes published, "type :list"
+    assert_includes unpublished, "type :list"
+    assert_includes published, "KitQuery.published_kits"
+    assert_includes unpublished, "KitQuery.unpublished_kits"
 
     query = File.read(File.expand_path("../lib/recording_studio_presskits/kit_query.rb", __dir__))
     assert_includes query, "recording_studio_trashable_active"
     assert_includes query, "def live_children"
     assert_includes query, "def live_child"
+    assert_includes query, "def published_kits"
+    assert_includes query, "PressKit.indexable"
+    assert_includes query, "def unpublished_kits"
   end
 
   def test_user_slice_uses_segmented_buttons_and_picker